+++ malis/malis_cpp.h
@@ -5,7 +5,7 @@ void connected_components_cpp(const int nVert,
                const int nEdge, const int* node1, const int* node2, const int* edgeWeight,
                int* seg);
 
-void malis_loss_weights_cpp(const int nVert, const int* seg,
+void malis_loss_weights_cpp(const int nVert, const int64_t* seg,
                const int nEdge, const int* node1, const int* node2, const float* edgeWeight,
                uint64_t* nPosPairPerEdge, uint64_t* nNegPairPerEdge,
 			   bool ignore_background = false,
